refactor: 점수 출력 관련 로직 Status 로 분리
