+++ src/test/java/chess/domain/BoardTest.java
@@ -28,19 +28,6 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class BoardTest {
 
-    private static ByteArrayOutputStream outputMessage;
-
-    @BeforeEach
-    void setUpStreams() {
-        outputMessage = new ByteArrayOutputStream();
-        System.setOut(new PrintStream(outputMessage));
-    }
-
-    @AfterEach
-    void restoresStreams() {
-        System.setOut(System.out);
-    }
-
     @Test
     @DisplayName("초기화된 체스판인지 확인")
     void isInitializedTrue() {
@@ -140,32 +127,5 @@ public class BoardTest {
         pieceExistBoard.put(Position.from("a2"), rook);
 
         return new Board(() -> pieceExistBoard);
-    }
-
-    @Test
-    @DisplayName("게임 점수 계산")
-    void showStatus() {
-        Map<Position, Piece> pieces = new HashMap<>();
-        pieces.put(Position.from("a1"), new Queen(Color.WHITE));
-        pieces.put(Position.from("a2"), new Rook(Color.BLACK));
-
-        Board board = new Board(() -> pieces);
-
-        board.showStatus(Output::printScore);
-        assertEquals("흰색의 점수는 9.0점 입니다.\n검은색의 점수는 5.0점 입니다.\n", outputMessage.toString());
-    }
-
-    @Test
-    @DisplayName("폰이 같은 x 에 존재할 경우의 게임 점수 계산")
-    void showStatusPawnScore0_5() {
-        Map<Position, Piece> pieces = new HashMap<>();
-        pieces.put(Position.from("a1"), new Pawn(Color.WHITE));
-        pieces.put(Position.from("a2"), new Pawn(Color.WHITE));
-        pieces.put(Position.from("a3"), new Pawn(Color.BLACK));
-
-        Board board = new Board(() -> pieces);
-
-        board.showStatus(Output::printScore);
-        assertEquals("흰색의 점수는 1.0점 입니다.\n검은색의 점수는 1.0점 입니다.\n", outputMessage.toString());
     }
 }
